+++ packages/camera/camera/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.8.1+5
+
+* Make sure the `setFocusPoint` and `setExposurePoint` coordinates work correctly in all orientations on iOS (instead of only in portrait mode). 
+
 ## 0.8.1+4
 
 * Silenced warnings that may occur during build when using a very